requirements.txt
DBUtils==1.3
futures==3.1.1
numpy==1.21.2
pandas==1.3.3
psycopg2-binary==2.9.1
PyMySQL==1.0.2
requests==2.26.0
SQLAlchemy==1.4.23
python-dotenv==0.19.0
pymongo==3.12.0
boto3==1.18.42